refactor: relocated files
arranged directory structure by meaning
