+++ todo/src/main/java/com/ssafy/sandbox/todo/domain/Todo.java
@@ -1,4 +1,4 @@
-package com.ssafy.sandbox.todo.model.domain;
+package com.ssafy.sandbox.todo.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
