+++ todo/src/main/java/com/ssafy/sandbox/todo/repository/TodoRepository.java
@@ -1,7 +1,7 @@
-package com.ssafy.sandbox.todo.model.repository;
+package com.ssafy.sandbox.todo.repository;
 
-import com.ssafy.sandbox.todo.model.domain.Todo;
-import com.ssafy.sandbox.todo.model.domain.TodoVO;
+import com.ssafy.sandbox.todo.domain.Todo;
+import com.ssafy.sandbox.todo.dto.TodoVO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
